Move framework/upgrade_util.go to framework/lifecycle/upgrade.go
